requirements.txt
Django==1.9
feedparser==5.1.3
south==0.7.6
pytz==2013b
psycopg2==2.6.1
djangorestframework==3.3.2
requests==2.8.1